dincli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dincli"
version = "0.1.0"
description = "Command-line interface for Decentralized Intelligence Network"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "DIN Dev Team", email = "contact@din.network" }
]
license = { text = "MIT" }
keywords = ["blockchain", "federated-learning", "privacy", "cli", "fastapi", "DIN"]

dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "python-dotenv>=1.0.0",
  "web3>=6.0.0",
  "requests>=2.31.0"
]

[project.scripts]
dincli = "cli.main:app"

[tool.setuptools.packages.find]
where = ["fastapi/pyapp"]
